+++ modules/CRM/Roundcube/RC/config/main.inc.php
@@ -2,13 +2,16 @@
 $d = getcwd();
 defined("_VALID_ACCESS") || define("_VALID_ACCESS", true);
 chdir(dirname(dirname(dirname(dirname(dirname(dirname(__FILE__)))))));
+define('SET_SESSION', false); // do not load EPESI session
 require_once('include/data_dir.php');
 require_once('include/config.php');
 require_once('include/database.php');
+require_once('include/session.php'); // load to get class in runtime
 global $E_SESSION,$E_SESSION_ID;
 $E_SESSION_ID = $_COOKIE[session_name()];
 if(!$E_SESSION_ID)
     $E_SESSION_ID = $_REQUEST[session_name()];
+$E_SESSION_ID = DBSession::truncated_session_id($E_SESSION_ID);
 $E_SESSION = DB::GetOne('SELECT data FROM session WHERE name = %s', array($E_SESSION_ID));
 if($E_SESSION)
     $E_SESSION = unserialize($E_SESSION);