+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ CURRENT
 -------
 
 - Fix Base/Print filename suffix
+- Fix not working RoundCube due to not loaded DBSession class
 
 
 RELEASE 1.5.6-20140303